% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapR.R
\name{add_points}
\alias{add_points}
\title{Add points to map}
\usage{
add_points(
  map,
  points_data,
  size = 0.3,
  alpha = 1,
  shape = 21,
  color = "orange",
  palette = "brewer.dark2"
)
}
\arguments{
\item{map}{Map object returned from plot_map(), plot_empty_map() or add_points()}

\item{points_data}{data frame containing LONG and LAT of each point}

\item{size}{Point plotting size}

\item{alpha}{Point alpha (transparency) value. Default = 1 (Solid)}

\item{shape}{Point shape. Default = 21 (Circle).}

\item{color}{Point plotting colour (Set to category column name for variable colour plotting)}

\item{palette}{Colour palette}
}
\description{
Add points to map
}
