% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapR.R
\name{add_radii}
\alias{add_radii}
\title{Add circles with specified radii to map}
\usage{
add_radii(
  map,
  points_data,
  radii,
  units = "km",
  alpha = 0.5,
  color = "orange",
  palette = "Dark2"
)
}
\arguments{
\item{map}{Map object returned from plot_map(), plot_empty_map() or add_points()}

\item{points_data}{data frame containing LONG and LAT of each point}

\item{radii}{radii of the circles}

\item{alpha}{Point alpha (transparency) value. Default = 1 (Solid)}

\item{color}{Point plotting colour (Set to category column name for variable colour plotting)}

\item{palette}{Colour palette}

\item{unit}{units for radius ("km", "m", "miles")}

\item{shape}{Point shape. Default = 21 (Circle).}
}
\description{
Add circles with specified radii to map
}
