% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GPcoverter.R
\name{convert_GP_data}
\alias{convert_GP_data}
\title{Title}
\usage{
convert_GP_data(
  data,
  GP_code_header,
  value_header,
  to = "Ward",
  norm_header = "None",
  norm_output_per = 100
)
}
\arguments{
\item{data}{data frame containing GP-level data}

\item{GP_code_header}{Column header for GP Code (e.g. M81062)}

\item{value_header}{Column header for variable of interest}

\item{to}{Area type to convert to (Ward/Constituency/Locality)}

\item{norm_header}{Column header for normalisation values}

\item{norm_output_per}{Normalisation level (e.g. per 100, per 1000)}
}
\description{
Title
}
