% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapR.R
\name{get_long_lat}
\alias{get_long_lat}
\title{Gets the longitude and latitude from West Midlands postcodes if LONG and LAT
columns are not already present}
\usage{
get_long_lat(points_data)
}
\arguments{
\item{points_data}{Data frame containing postcodes and/or longitude and
latitude data}
}
\description{
Columns must be labelled exactly as "Postcode" or "LONG" and "LAT".
}
\details{
Function only works with West Midlands Postcodes.
}
