% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapR.R
\name{get_points_shape}
\alias{get_points_shape}
\title{Converts data frame with longitude and latitude}
\usage{
get_points_shape(points_data)
}
\arguments{
\item{points_data}{Data frame containing longitude and latitude data}
}
\description{
Longitude and latitude columns names must be exactly "LONG" and "LAT"
}
