% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapR.R
\name{get_radii_shapes}
\alias{get_radii_shapes}
\title{Create radii shape object}
\usage{
get_radii_shapes(points_shape, radii, units = "km")
}
\arguments{
\item{radii}{radii of the circles}

\item{units}{units of the radii (Must be "km", "m", or "miles")
Other units might be okay, but they haven't been tested
TODO: Test these 3}

\item{points_data}{data frame containing LONG and LAT of each point}
}
\description{
Creates shape file for circles centered on the given coordinates with the
defined radii
}
