% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{load_data}
\alias{load_data}
\title{Load data file}
\usage{
load_data(file_path = "output_data.xlsx", sheet = 1)
}
\arguments{
\item{sheet}{Sheet name/number (only for xls/xlsx files)}

\item{file_name}{Path to data file (.xls, .xlsx or .csv)}
}
\value{
data frame containing loaded data
}
\description{
Load data file
}
