% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapR.R
\name{plot_empty_map}
\alias{plot_empty_map}
\title{Plot Empty Map}
\usage{
plot_empty_map(
  map_title = "",
  area_name = "BSol",
  map_type = "Ward",
  pallet = "Blues",
  const_lines = "None",
  const_names = "None",
  locality_lines = "None",
  locality_names = "None",
  compass = TRUE,
  credits = "default",
  credits_size = 0.6,
  verbose = FALSE
)
}
\arguments{
\item{map_title}{Title for the map}

\item{area_name}{Name of area to be plotted: BSol, Birmingham, or Solihull}

\item{map_type}{Map geography type: Constituency, Ward, LSOA21, etc}

\item{pallet}{Colour pallet}

\item{const_lines}{Include constituency lines: TRUE/FALSE}

\item{const_names}{Include constituency names: TRUE/FALSE}

\item{locality_lines}{Include locality lines: TRUE/FALSE}

\item{locality_names}{Include locality names: TRUE/FALSE}

\item{compass}{Include compass: TRUE/FALSE}

\item{credits}{Credit text to be shown at bottom of the map}

\item{credits_size}{Credit size}

\item{verbose}{Print debugging text}
}
\description{
Plot Empty Map
}
