% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapR.R
\name{plot_map}
\alias{plot_map}
\title{Plot map}
\usage{
plot_map(
  data,
  value_header,
  map_type,
  area_name = "BSol",
  map_title = "",
  fill_missing = NA,
  palette = "Blues",
  style = "pretty",
  breaks = NULL,
  labels = NULL,
  const_lines = "None",
  const_names = "None",
  locality_lines = "None",
  locality_names = "None",
  compass = TRUE,
  credits = "default",
  credits_size = 0.6,
  verbose = FALSE
)
}
\arguments{
\item{data}{Data frame containing area IDs and plot value}

\item{value_header}{Header name for the value to be plotted}

\item{map_type}{Map geography type: Constituency, Ward, LSOA21, etc}

\item{area_name}{Name of area to be plotted: BSol, Birmingham, or Solihull}

\item{map_title}{Title for the map}

\item{fill_missing}{Fill missing values (default = NA)}

\item{style}{Colour style: pretty/fixed}

\item{breaks}{Value plotting range}

\item{labels}{Custom value labels}

\item{const_lines}{Include constituency lines: TRUE/FALSE}

\item{const_names}{Include constituency names: TRUE/FALSE}

\item{locality_lines}{Include locality lines: TRUE/FALSE}

\item{locality_names}{Include locality names: TRUE/FALSE}

\item{compass}{Include compass: TRUE/FALSE}

\item{credits}{Credit text to be shown at bottom of the map}

\item{credits_size}{Credit size}

\item{verbose}{Print debugging text}

\item{paletteColour}{palette}
}
\description{
Plot map
}
