% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{save_data_file}
\alias{save_data_file}
\title{Save data file}
\usage{
save_data_file(data, save_path = "output_data.xlsx")
}
\arguments{
\item{data}{Data frame to be saved}

\item{save_path}{Save file name and path}
}
\description{
Save data file
}
