% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapR.R
\name{save_map}
\alias{save_map}
\title{Save map object}
\usage{
save_map(map, save_name = "new_map.png", height = 5, width = 5)
}
\arguments{
\item{map}{Map object returned from plot_map(), plot_empty_map() or add_points()}

\item{save_name}{File save name}

\item{height}{Map figure height (inches)}

\item{width}{Map figure width (inches)}
}
\description{
Save map object
}
